Added casting variable from validate_number_u series function.

+++ src/calendar.c
@@ -26,7 +26,7 @@ u32 validate_number_u32(u32 uv)
 		uv = (u32) (0xFFFFFFFF - (uv & 0xFFFFFFFF) + 1);
 	return uv;
 	*/
-	return shift_bits(uv, none);
+	return (u32) shift_bits(uv, none);
 }
 #endif
 
@@ -38,7 +38,7 @@ u16 validate_number_u16(u16 uv)
 		uv = (u16) (0x0000FFFF - (uv & 0x0000FFFF) + 1);
 	return uv;
 	*/
-	return shift_bits(uv, word);
+	return (u16) shift_bits(uv, word);
 }
 
 static
@@ -49,7 +49,7 @@ u8 validate_number_u8(u8 uv)
 		uv = (u8)  (0x000000FF - (uv & 0x000000FF) + 1);
 	return uv;
 	*/
-	return shift_bits(uv, dword);
+	return (u8) shift_bits(uv, dword);
 }
 
 /**
